requirements.txt
wheel
python-dotenv
flask
pytest
google-cloud-storage
